requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pymongo==4.5.0
PyMuPDF==1.23.7
Pillow==10.0.1
python-pptx==0.6.21
azure-ai-documentintelligence==1.0.0
openai==1.12.0
python-multipart==0.0.6